% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_quantized_curve.R
\name{add_quantized_curve}
\alias{add_quantized_curve}
\title{Add another quantized curve to quantized curve plot}
\usage{
add_quantized_curve(
  betas = 1,
  intercept = 0,
  nsteps = 5,
  link = plogis,
  steps,
  center = TRUE,
  curve_color = "black",
  quant_color = "red",
  curve_linesize = 1,
  quant_linesize = 0.9,
  quant_pointsize = 3,
  hide_curve = FALSE
)
}
\arguments{
\item{betas}{Numeric vector of polynomial term coefficients}

\item{intercept}{Intercept to use}

\item{nsteps}{Number of points to use in the quantized curve}

\item{link}{Link function to use, defaults to \code{plogis}}

\item{steps}{Specific steps to use (i.e., x-values to use)}

\item{center}{Logical, if \code{steps} is specified, whether to center the values.
Defaults to TRUE.}

\item{curve_color}{Color to use for the identification curve}

\item{quant_color}{Color to use for the quantized curve}

\item{curve_linesize}{Line width for the identification curve, default 1}

\item{quant_linesize}{Line width for the quantized curve, default .9}

\item{quant_pointsize}{Point size for the quantized curve, default 3}

\item{hide_curve}{Logical, defaults to FALSE, whether to hide the identification
curve and ONLY show the quantized curve.}
}
\value{
ggplot layers for an identification curve and its quantized curve
}
\description{
Same usage as \code{plot_quantized_curve} but only returns the ggplot layers
for the curve and quantized curve. This can be added to a previously made
plot from \code{plot_quantized_curve} to show two curves, or two different
quantizations.
}
\examples{

plot_quantized_curve(c(1,0,1), 0, 5) +
add_quantized_curve(c(2,0,2), 0, 5, curve_color = "blue")
}
