% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bivariate_surface.R
\name{add_quantized_surface}
\alias{add_quantized_surface}
\title{Add quantized bivariate surface}
\usage{
add_quantized_surface(
  plotly_fig,
  betas1 = c(1, 0.25, -0.01),
  betas2 = c(2, -0.25, -0.01),
  intercept1 = 0,
  intercept2 = 0,
  domain1 = c(-3, 3),
  domain2 = c(-3, 3),
  label1 = "X1",
  label2 = "X2",
  labeldv = "Probability",
  link = plogis,
  smooth_by = 0.1,
  nsteps = 5,
  steps1,
  steps2,
  point_colors = c("blue", "gray", "red"),
  color_by = 1L,
  show_quantized_points = TRUE
)
}
\arguments{
\item{plotly_fig}{Plotly figure to add to, result of \code{plot_bivariate_surface}}

\item{betas1}{Numeric vector of polynomial term coefficients for polynomial 1}

\item{betas2}{Numeric vector of polynomial term coefficients for polynomial 2}

\item{intercept1}{Numeric intercept for polynomial 1}

\item{intercept2}{Numeric intercept for polynomial 2}

\item{domain1}{Numeric vector of length 2 for start and end points of the domain
of polynomial 1, defaults to \code{c(-3, 3)}}

\item{domain2}{Numeric vector of length 2 for start and end points of the domain
of polynomial 2, defaults to \code{c(-3, 3)}}

\item{label1}{Label to use for polynomial 1's axis, defaults to \code{X1}}

\item{label2}{Label to use for polynomial 2's axis, defaults to \code{X2}}

\item{labeldv}{Label to use for the vertical dependent variable between 0 and
1, defaults to \code{Probability}}

\item{link}{Link function to use, defaults to \code{plogis}}

\item{smooth_by}{Numeric value to use to interpolate steps along the polynomial
axes, defaults to \code{.1} (i.e., \code{seq(-3, 3, by=.1)})}

\item{nsteps}{Number of steps to interpolate between for quantized points,
defaults to \code{5}}

\item{steps1}{Numeric vector of step values for polynomial 1}

\item{steps2}{Numeric vector of step values for polynomial 2}

\item{point_colors}{Character vector of color names to create a gradient
between, defaults to \code{c('blue','gray','red')}}

\item{color_by}{Either 1 to color the points by each step along \code{X1} or 2
to color by each step along \code{X2}. Bivariate gradient currently not supported.}

\item{show_quantized_points}{Holdover from \code{plot_bivariate_surface}, does
nothing here}
}
\value{
Plotly figure with a point trace added
}
\description{
Add quantized points to a given surface
}
