% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animate_interpolated_curves.R
\name{animate_interpolated_curves}
\alias{animate_interpolated_curves}
\title{Animate between interpolated curves}
\usage{
animate_interpolated_curves(
  betas = list(c(0, 0, 0), c(2, 0, 0), c(2, 0.5, 0), c(2, 0.5, -0.15)),
  nsteps_between = 10,
  intercept = 0,
  link = plogis,
  colors = c("black", "blue", "gray", "red"),
  linesize = 1,
  animations = list(gganimate::enter_fade(), gganimate::exit_disappear())
)
}
\arguments{
\item{betas}{List of numeric vectors (all of same length)}

\item{nsteps_between}{Number of steps to interpolate between, integer}

\item{intercept}{Intercept to use}

\item{link}{link function, defaults to \code{plogis}}

\item{colors}{Colors to use for the gradient, recommended to provide as many
colors as there are curves to interpolate between}

\item{linesize}{Size for the lines, defaults to 1}

\item{animations}{\code{gganimate} \code{exit_} and \code{enter_} functions to use, as a list}
}
\value{
Animated ggplot
}
\description{
Given a list of polynomials, animate how the identification function changes
}
