% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animate_interpolated_curves.R
\name{.format_polynomial_strings}
\alias{.format_polynomial_strings}
\title{Write out polynomial}
\usage{
.format_polynomial_strings(curve_df, intercept = 0)
}
\arguments{
\item{curve_df}{Data frame defining the curves}

\item{intercept}{Intercept to use}
}
\value{
Character vector of polynomials
}
\description{
Write out polynomial
}
