% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_values_factor.R
\name{.get_values_factory}
\alias{.get_values_factory}
\title{Curve point calculation function factory}
\usage{
.get_values_factory(betas, intercept = 0, link)
}
\arguments{
\item{betas}{Numeric vector of polynomial term coefficients}

\item{intercept}{Intercept for the polynomial, must be length 1}

\item{link}{Link function to use}
}
\value{
Function to calculate probabilities of given x values
}
\description{
Creates a function that can be used to calculate values of a polynomial
with a given link function.
}
