% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate_polynomials.R
\name{.interpolate_polynomials}
\alias{.interpolate_polynomials}
\title{Interpolate polynomial terms}
\usage{
.interpolate_polynomials(from = c(0, 0), to = c(5, 2), nsteps = 6)
}
\arguments{
\item{from}{Numeric vector of polynomial terms to start at}

\item{to}{Numeric vector of polynomial terms to end at}

\item{nsteps}{Number of steps to interpolate between}
}
\value{
\code{data.frame} of numeric values where columns correspond to polynomial
terms and rows correspond to the interpolated values
}
\description{
This is a helper function to interpolate between two endpoints along each
polynomial term coefficient.
}
