% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_id_derivative.R
\name{.make_derivative_ylabs}
\alias{.make_derivative_ylabs}
\title{Calculate y-axis breaks and labels}
\usage{
.make_derivative_ylabs(curves_df)
}
\arguments{
\item{curves_df}{Dataframe with the y and yprime values}
}
\value{
A list with breaks, labels, and hjust values
}
\description{
Helper to get the y-axis breaks, labels, and justification
}
