% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_id_curve.R
\name{plot_id_curve}
\alias{plot_id_curve}
\title{Plot an identification curve}
\usage{
plot_id_curve(
  betas,
  intercepts,
  colors = c("blue", "gray", "red"),
  link = plogis,
  linesize = 1
)
}
\arguments{
\item{betas}{Numeric vector of polynomial term coefficients}

\item{intercepts}{Intercepts to use when plotting the same polynomial terms
with different intercepts}

\item{colors}{Discrete points for a continuous color gradient. Defaults
to blue, gray, red (ascending from blue towards red)}

\item{link}{Link function to use, defaults to plogis (i.e., logit link)}

\item{linesize}{Width of the lines to use. Defaults to 1.}
}
\value{
A ggplot of the plotted polynomials
}
\description{
Plots an identification curve using the given polynomial terms and
intercepts. If no intercepts are given, a preset range of -3 to 3 will be
plotted. A reference curve of link(1x+0) is shown in a dashed line.
}
\details{
Link function should be of the \verb{p*} family of functions, such as
\code{pnorm}, \code{plogis}, \code{plaplace} etc.
}
\examples{

plot_id_curve(c(2, .2, -.3), 0)

}
