% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_id_derivative.R
\name{plot_id_derivative}
\alias{plot_id_derivative}
\title{Plot identification curve with rate of change curve}
\usage{
plot_id_derivative(
  betas = c(1, 0, -10),
  intercept = 0,
  link = plogis,
  domain = c(-3, 3),
  smooth_by = 0.1,
  curve_color = "black",
  deriv_color = "royalblue4"
)
}
\arguments{
\item{betas}{Numeric vector of polynomial term coefficients}

\item{intercept}{Intercept to use}

\item{link}{Link function, defaults to \code{plogis}}

\item{domain}{Domain of the x-axis for the function, default \code{c(-3, 3)}}

\item{smooth_by}{Precision of the numeric sequence between the domain ends}

\item{curve_color}{Color of the identification curve and left y-axis,
default \code{'black'}}

\item{deriv_color}{Color of the rate of change curve and right y-axis,
default \verb{'royalblue4}}
}
\value{
A ggplot with the identification curve and rate of change curve
}
\description{
Plots the identification curve with given polynomial term coefficients
and intercept along with the first derivative approximation from
\code{numDeriv::grad}, reflecting the rate of change in the identification curve.
}
\details{
Take note of the double y-axis.
}
