% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_quantized_curve.R
\name{plot_quantized_curve}
\alias{plot_quantized_curve}
\title{Plot quantized identification curve}
\usage{
plot_quantized_curve(
  betas = 1,
  intercept = 0,
  nsteps = 5,
  link = plogis,
  steps,
  center = TRUE,
  expand = TRUE,
  curve_color = "black",
  quant_color = "red",
  curve_linesize = 1,
  quant_linesize = 0.9,
  quant_pointsize = 3,
  hide_curve = FALSE
)
}
\arguments{
\item{betas}{Numeric vector of polynomial term coefficients}

\item{intercept}{Intercept to use}

\item{nsteps}{Number of points to use in the quantized curve}

\item{link}{Link function to use, defaults to \code{plogis}}

\item{steps}{Specific steps to use (i.e., x-values to use)}

\item{center}{Logical, if \code{steps} is specified, whether to center the values.
Defaults to TRUE.}

\item{expand}{Logical, whether to expand the x-axis a bit to show more of
the identification curve. Defaults to TRUE, if FALSE it will essentially zoom
in on the quantized portion.}

\item{curve_color}{Color to use for the identification curve}

\item{quant_color}{Color to use for the quantized curve}

\item{curve_linesize}{Line width for the identification curve, default 1}

\item{quant_linesize}{Line width for the quantized curve, default .9}

\item{quant_pointsize}{Point size for the quantized curve, default 3}

\item{hide_curve}{Logical, defaults to FALSE, whether to hide the identification
curve and ONLY show the quantized curve.}
}
\value{
A ggplot of the identification curve with the quantized curve on top
of it.
}
\description{
A quantized curve is an approximation of a continuous curve using a smaller
discrete set of points with straight-line interpolations between the points.
}
\details{
Plots an identification curve quantized to a given a number of evenly-spaced
steps or a given vector of step locations (that don't have to be evenly-spaced).
}
\examples{

plot_quantized_curve(c(1, 4, -.1), intercept = 0, nsteps = 5)
}
