% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_interpolations.R
\name{table_interpolations}
\alias{table_interpolations}
\title{Table of interpolated values}
\usage{
table_interpolations(
  from = c(0, 0),
  to = c(5, 2),
  nsteps = 6L,
  precision = 2L,
  colors = c("blue", "gray", "red")
)
}
\arguments{
\item{from}{Numeric vector of polynomial terms to start at}

\item{to}{Numeric vector of polynomial terms to end at}

\item{nsteps}{Integer number of steps to interpolate between}

\item{precision}{Rounding precision for table, defaults to 2}

\item{colors}{Discrete points for a continuous color gradient. Defaults
to blue, gray, red (ascending from blue towards red)}
}
\value{
A gt table of the interpolated polynomial coefficients with colored
labels following the gradient defined by \code{colors}
}
\description{
Used to create a kind of legend for plots from \code{plot_interpolated_curves}.
Note that the colors used should be the same in order to match the plot.
}
